Major additions.
It seems the workshop website updated the submission instructions.
References are now excluded from the page limit and thus I had an extra
page of space that could be utilized.  I decided to add a few paragraphs
on how Lina relates to the existing work on using automated theorem
provers to generate and analyize attack trees.